requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
python-dotenv==1.0.0
pymysql==1.1.0